+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Bugs Fixed
 
 * [#216](https://github.com/clojure-emacs/orchard/issues/216): Fix some pagination issues with inspector navigation.
-  * `down` fn goes beyond current page. `next-sibling` and `previous-sibling` fns work in all pages.
+  * `orchard.inspect/down` fn is now able to go beyond the current page. `next-sibling` and `previous-sibling` fns work in all pages.
 
 ## 0.23.0 (2024-03-03)